+++ cleosim/__init__.py
@@ -4,12 +4,13 @@ import cleosim.opto
 import cleosim.coordinates
 import cleosim.stimulators
 import cleosim.recorders
+import cleosim.processing
 import cleosim.utilities
-# bring nested modules up to second level
-from cleosim.base import *
-from cleosim.processing import controllers, delays, observers
 
-from ._version import get_versions
-
-__version__ = get_versions()["version"]
-del get_versions
+from cleosim.base import (
+    CLSimulator,
+    Recorder,
+    Stimulator,
+    InterfaceDevice,
+    ProcessingLoop,
+)
