+++ cleosim/electrodes/__init__.py
@@ -1,6 +1,13 @@
-
 """Electrodes  module"""
-from cleosim.electrodes import *
-from cleosim.electrodes.lfp import *
-from cleosim.electrodes.spiking import *
-from cleosim.electrodes.probes import *
+from cleosim.electrodes.lfp import TKLFPSignal
+from cleosim.electrodes.spiking import MultiUnitSpiking, SortedSpiking, Spiking
+from cleosim.electrodes.probes import (
+    Probe,
+    Signal,
+    concat_coords,
+    linear_shank_coords,
+    tetrode_shank_coords,
+    poly2_shank_coords,
+    poly3_shank_coords,
+    tile_coords,
+)
